Streamline main module and add proper connection handling
- Please note that it's not recommended to catch top-level Exception.
In case of sqlite3, the Error is suitable candidate for the job.

- In this case, it's probably better to raise the exception again after
logging it, as we'd rather want the app to crash than let the users use
it without the properly loaded data or, even worse, without the ability
to save their new entries.

- Failure to establish a DB connection is logged, but it would be more
user-friendly to also inform about this situation via a simple dialogue
window allowing the user to acknowledge the situation before the app
stops itself.

- Commenting your code is important. However, when you decide to write
a comment, use it to describe WHY are you doing that particular thing,
not WHAT is that particular thing doing, as the latter should be obvious
from the code itself, whereas the former may be forgotten even by the
code authors themselves.

- It is better to close the DB connection even if the app crashes.

- Sometimes it's better to use just plain, simple approach. The configuration
"parsing" in the main function seemed rather unnecessary as we can use format,
which is friendly both to user and code out-of-the-box.