Showcase of Enum usage
Enums are especially fit for the cases when one has a limited set of values
which logically represent particular group / type. One can view them as a set
of constants of a common type. One of the advantages of enums is that, unlike
literals, the enums can be easily refactored and their usage can be reliably
traced across the whole codebase. Other advantage is the type control when
you need to pass an enum value as parameter.