Simplify button callback
Demonstrates, how to simplify implementation of the button_clicked callback method
as well as the construction of ttk.Button grid.

Please note the "lambda d=day" instead of simple "lambda d" is necessary to "capture"
immediate value of "d" at the time the lambda is assigned as the command of particular
button. Otherwise, a late binding would be used resulting in all "d" values being
identical to value of "day_" after the last iteration.