Showcase list comprehensions by building week list
- List comprehensions are recommended way of building a list (generally
preferred over for loop with .append)

- Reusing list (month_matrix from from cal.monthcalendar) for other purposes
(overwriting some elements with Day instance) is generally not recommended.
Unless there's a performance reason (e.g. very huge lists or items are objects
which are costly to create / fetch from DB / etc), it's more common to construct
a new list by "mapping" items of old list to the new ones

In this case there are two list comprehensions

1) nested one which converts items of one week to Days
[day(year_, month_, day_) if day_ else None for day_ in week]

2) outer one, which just collects individual Weeks into a list of weeks
weeks = [
  [ ...inner...]
  for week in month_calendar
]