Drop unnecessary widget references
It's not necessary to keep reference to every component (widget). Only those
which will be interacted with directly should be saved. Labels can be updated
via their bound tk-variables, buttons execute the command callback themselves.
In the context of the MainPage.populate_frame, the only reference which should
be kept for later interaction is the entry_textbox.