Update loggin according to recommended best practices
While f-strings will work perfectly fine, it's a common practice to
use "%" style formatters with logging. The reason is simple - the
f-strings are evaluated everytime, whereas the "%" only if current
loglevel allows the log information to be written. Hence "%" saves
some CPU cycles. This isn't significant for simple desktop app, but
can make a big difference for server apps with tens of thousands of
requests per minute.